Did some stuff that doesn't really pan out

+++ src/memory/memory.rs
@@ -2,8 +2,8 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 pub trait Memory {
-  fn read(&self, address: usize) -> u8;
-  fn write(&mut self, address: usize, value: u8);
+  fn read(&self, address: u16) -> u8;
+  fn write(&mut self, address: u16, value: u8);
 }
 
 pub type MemoryRef = Rc<RefCell<dyn Memory>>;
